+++ packages/typespec-go/src/tcgcadapter/clients.ts
@@ -70,10 +70,6 @@ export class clientAdapter {
       clientName = parent.name.substring(0, parent.name.length - 6) + clientName;
     }
 
-    if (!clientName.match(/Client$/)) {
-      clientName += 'Client';
-    }
-
     const docs: go.Docs = {
       summary: sdkClient.summary,
       description: sdkClient.doc,
@@ -412,7 +408,7 @@ export class clientAdapter {
             throw new AdapterError('UnsupportedTsp', `unsupported spread param content type ${contentType}`, opParam.__raw?.node ?? NoTarget);
         }
       } else {
-        adaptedParam = this.adaptMethodParameter(opParam);
+        adaptedParam = this.adaptMethodParameter(opParam, method.httpMethod);
       }
 
       adaptedParam.docs.summary = param.summary;
@@ -437,7 +433,7 @@ export class clientAdapter {
     // look for them in the operation parameters.
     for (const opParam of allOpParams) {
       if (opParam.onClient) {
-        const adaptedParam = this.adaptMethodParameter(opParam);
+        const adaptedParam = this.adaptMethodParameter(opParam, method.httpMethod);
         adaptedParam.docs.summary = opParam.summary;
         adaptedParam.docs.description = opParam.doc;
         method.parameters.unshift(adaptedParam);
@@ -480,7 +476,7 @@ export class clientAdapter {
     return contentType;
   }
 
-  private adaptMethodParameter(param: tcgc.SdkBodyParameter | tcgc.SdkHeaderParameter | tcgc.SdkPathParameter | tcgc.SdkQueryParameter | tcgc.SdkCookieParameter): go.MethodParameter {
+  private adaptMethodParameter(param: tcgc.SdkBodyParameter | tcgc.SdkHeaderParameter | tcgc.SdkPathParameter | tcgc.SdkQueryParameter | tcgc.SdkCookieParameter, verb: go.HTTPMethod): go.MethodParameter {
     if (param.isApiVersionParam && param.clientDefaultValue) {
       // we emit the api version param inline as a literal, never as a param.
       // the ClientOptions.APIVersion setting is used to change the version.
@@ -513,7 +509,10 @@ export class clientAdapter {
     }
 
     let adaptedParam: go.MethodParameter;
-    const paramStyle = this.adaptParameterStyle(param);
+    let paramStyle = this.adaptParameterStyle(param);
+    if (param.kind === 'body' && (verb === 'patch' || verb === 'put')) {
+      paramStyle = 'required';
+    }
     const paramName = getEscapedReservedName(ensureNameCase(param.name, paramStyle === 'required'), 'Param');
     const byVal = isTypePassedByValue(param.type);